added water hole in tilemaps to avoid bug
theres a bug in the current new tilemaps where the third from the middle-bottom tile will not highlight while showing a unit's range.

to avoid showing this in the standup, I added a waterhole in the middle of the two maps. will need to fix in the future! just providing a work-around option for now